:hammer: refactor: Modify member info dto
