+++ types/channel/response/channel-member-info-response.interface.ts
@@ -1,11 +1,4 @@
-export type Role = 'owner' | 'admin' | 'member';
-
-export interface MemberInfo {
-  id: number;
-  nickname: string;
-  image?: string;
-  role: Role;
-}
+import { MemberInfo } from '../socket';
 
 export interface ChannelMemberInfoResponse {
   players: MemberInfo[];
