+++ types/channel/socket/member-info.interface.ts
@@ -1,0 +1,8 @@
+export type Role = 'owner' | 'admin' | 'member';
+
+export interface MemberInfo {
+  userId: number;
+  nickname: string;
+  image?: string;
+  role: Role;
+}
