--- types/channel/socket/new-member.interface.ts
@@ -1,5 +1,0 @@
-export interface NewMember {
-  userId: number;
-  nickname: string;
-  image?: string;
-}
